+++ clang/include/clang/Sema/Sema.h
@@ -3522,6 +3522,10 @@ public:
   /// attribute.
   mutable llvm::DenseMap<const EnumDecl *, llvm::APInt> FlagBitsCache;
 
+  /// A cache of enumerator values for enums checked by -Wassign-enum.
+  mutable llvm::DenseMap<const EnumDecl *, llvm::SmallVector<llvm::APSInt, 64>>
+      AssignEnumCache;
+
   /// WeakUndeclaredIdentifiers - Identifiers contained in \#pragma weak before
   /// declared. Rare. May alias another identifier, declared or undeclared.
   ///